+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [Semantic Field] Support the sparse two phase processor for the semantic field.
 - [Stats] Add stats for agentic query and agentic query translator processor.
 - [Agentic Search] Adds validations and logging for agentic query
+- [Agentic Search] Convert agentic query translator processor to system-generated processor
 
 ### Bug Fixes